Allow connecting to wemo devices directly
This patch addresses two issues for me:
 1. I'd like to be able to talk to wemos directly over a VPN by IP
 2. The discovery mechanism is very unreliable at times

This patch allows this behavior by letting a user specify a device
manually by the URL needed to reach it. In order to make this work,
we need to implement the reconnect procedure differently because
we can't expect to be able to find it via SSDP.

Since probing should be more reliable (assuming the device is still
reachable at the same address) we try to probe first. If that does
not work, then we try SSDP rediscovery if we originally found the
wemo via that mechanism. If we didn't, then retrying SSDP doesn't
make any sense.

I've been using a variant of this in a local fork of pywemo for a
while and have found it to be much more reliable than depending on
discovery for initial setup and rediscovery.