Work arounds for what seem to be firmware bugs.
When we set brightness to 0 on a bulb that is on it will turn off
and non-zero on an off bulb will turn it back on. But on one of my
bulbs the 'onoff' state doesn't get updated.
Another problem is that a turned off bulb will first turn on to it's
last remembered brightness level and only then fade to the requested
level, so turning on to a dim setting would be a light that switches
on bright and then fades down instead of a slow fade-in.

To work around these,
- If 'sleepfader' is available, use that to turn off the bulb because
  it both dims to 0 over a time interval and updates onoff state.
- When fading from off to on we first force the dim level to the
  lowest non-off state (1), then send a onoff=ON event to update the
  on-off state and finally send the command to fade to the desired
  brightness.