Distinguish between connection and read timeouts when probing wemos
When probing for a wemo, we want to stop the probe immediately when
we get a connection timeout because it means the wemo didn't respond
at all (even with a RST on the invalid port). However, it appears
that sometimes wemos will get wedged such that they continue to
accept() on the old port, but won't respond. Our current logic treats
that the same as a connection timeout, when really we need to ignore
that case and continue to the next port.

This change breaks out ConnectTimeout from Timeout (which is really
just ReadTimeout). It bails on the former and continues on the latter
to give the desired behavior.