Roll back the use of the GetDeviceStatus call.
It turns out that what is reported by GetDeviceStatus is pretty unreliable.
All I had to do was turn on a poor quality lightbulb that has a high pitched
whine, it must be putting out quite a bit of RF noise too because although I
could still control my lights, they were reporting back as unavailable.

This in itself is not a big deal, but when changing the color values with
home-assistant while we are not getting valid status makes the bulbs flash
annoyingly.