Push temperature and color change requests onto a 'queue'.
Avoids duplicating kelvin -> mired code while still limiting the number
of packets we send on the zigbee bus.

The queue isn't really a sequence of commands to be executed, so it only
remembers the last colorcontrol and colortemperature value pushed. It is
also unspecified what bulbs do when both color and temperature are set
at the same time.