Fix for wemo link discovery issue.
Tonight my lights became inaccessible and I could not get reconnected
to the bridge. It turned out that we were getting an empty brightness
level for a light during discovery, which triggered an exception that
was silently caught and dropped.

Removing the exception handler made it immediately obvious what was
going on, maybe we should do without that exception handler for a bit?