Handle cases where the bridge doesn't know the brightness
If my bridge is power cycled then it has no idea what brightness the bulbs
are at, and appears to default to not providing that information on query.
If a client then tries to turn on the bulb but doesn't provide a level, we
get a KeyError. In the absence of any other information, assume that we want
the lights on full in that case.